feat: Исправил обновление данных профиля

+++ app/src/main/java/com/eltex/chat/data/models/profileinfo/ProfileInfoRequest.kt
@@ -16,5 +16,5 @@ data class ProfileInfoRequest(
     val statusConnection: String,
     val success: Boolean,
     val username: String,
-    val utcOffset: Int
+    val utcOffset: Float,
 )
